Re-organize files in `pipeline-settings`
