+++ services/orchest-webserver/client/src/pipeline-settings-view/common.ts
@@ -1,4 +1,4 @@
-import { ServiceTemplate } from "@/components/ServiceTemplatesDialog/content";
+import { ServiceTemplate } from "@/pipeline-settings-view/ServiceTemplatesDialog/content";
 import type { Json, PipelineJson, PipelineSettings, Service } from "@/types";
 import { hasValue } from "@orchest/lib-utils";
 import "codemirror/mode/javascript/javascript";
